Update ctypegen_comedi.py based on new ctypesgen
Using the newer version of ctypesgen that does not mangle the POINTER type
such that auto-byref calls get broken.  Using this change, calling comedi
functions becomes much more streamlined for users as one does not always
have to remember to use ctypes.byref.

The old behavior, when ctypes.byref was not used for pointer arguments, was
that a segfault was generally thrown.  This update fixes that issue.