python/examples: fix ai_mmap.py--now returns data
Fixes: 37a09df32c7a so that ai_mmap.py returns data.  Problem had been that
the inner loop was waiting on subdevice running status and was exiting
before the data was ready to be read.  This patch causes the inner loop to
wait on subdevice busy status which is true even if the analog acquisition
is finished--as long as there is actually data to read from the
(potentially) already finished acquisition.